requirements.txt
streamlit
pandas
numpy
requests
smartapi-python
pytz
pyotp
logzero
websocket-client